+++ aws-xray/src/main/java/io/opentelemetry/contrib/awsxray/AwsXrayRemoteSampler.java
@@ -5,6 +5,14 @@
 
 package io.opentelemetry.contrib.awsxray;
 
+import static java.util.concurrent.TimeUnit.NANOSECONDS;
+import static java.util.concurrent.TimeUnit.SECONDS;
+import static java.util.function.Function.identity;
+import static java.util.logging.Level.FINE;
+import static java.util.stream.Collectors.toList;
+import static java.util.stream.Collectors.toMap;
+import static java.util.stream.Collectors.toSet;
+
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.api.trace.SpanKind;
 import io.opentelemetry.context.Context;
@@ -29,17 +37,13 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.ThreadLocalRandom;
-import java.util.concurrent.TimeUnit;
-import java.util.function.Function;
-import java.util.logging.Level;
 import java.util.logging.Logger;
-import java.util.stream.Collectors;
 import javax.annotation.Nullable;
 
 /** Remote sampler that gets sampling configuration from AWS X-Ray. */
 public final class AwsXrayRemoteSampler implements Sampler, Closeable {
 
-  static final long DEFAULT_TARGET_INTERVAL_NANOS = TimeUnit.SECONDS.toNanos(10);
+  static final long DEFAULT_TARGET_INTERVAL_NANOS = SECONDS.toNanos(10);
 
   private static final Logger logger = Logger.getLogger(AwsXrayRemoteSampler.class.getName());
 
@@ -134,7 +138,7 @@ public final class AwsXrayRemoteSampler implements Sampler, Closeable {
                 initialSampler,
                 response.getSamplingRules().stream()
                     .map(SamplingRuleRecord::getRule)
-                    .collect(Collectors.toList())));
+                    .collect(toList())));
 
         previousRulesResponse = response;
         ScheduledFuture<?> existingFetchTargetsFuture = fetchTargetsFuture;
@@ -142,18 +146,17 @@ public final class AwsXrayRemoteSampler implements Sampler, Closeable {
           existingFetchTargetsFuture.cancel(false);
         }
         fetchTargetsFuture =
-            executor.schedule(
-                this::fetchTargets, DEFAULT_TARGET_INTERVAL_NANOS, TimeUnit.NANOSECONDS);
+            executor.schedule(this::fetchTargets, DEFAULT_TARGET_INTERVAL_NANOS, NANOSECONDS);
       }
     } catch (Throwable t) {
-      logger.log(Level.FINE, "Failed to update sampler", t);
+      logger.log(FINE, "Failed to update sampler", t);
     }
     scheduleSamplerUpdate();
   }
 
   private void scheduleSamplerUpdate() {
     long delay = pollingIntervalNanos + jitterNanos.next();
-    pollFuture = executor.schedule(this::getAndUpdateSampler, delay, TimeUnit.NANOSECONDS);
+    pollFuture = executor.schedule(this::getAndUpdateSampler, delay, NANOSECONDS);
   }
 
   /**
@@ -168,7 +171,7 @@ public final class AwsXrayRemoteSampler implements Sampler, Closeable {
     if (pollFuture == null) {
       return null;
     }
-    return Duration.ofNanos(pollFuture.getDelay(TimeUnit.NANOSECONDS));
+    return Duration.ofNanos(pollFuture.getDelay(NANOSECONDS));
   }
 
   private void fetchTargets() {
@@ -181,28 +184,25 @@ public final class AwsXrayRemoteSampler implements Sampler, Closeable {
       Date now = Date.from(Instant.ofEpochSecond(0, clock.now()));
       List<SamplingStatisticsDocument> statistics = xrayRulesSampler.snapshot(now);
       Set<String> requestedTargetRuleNames =
-          statistics.stream()
-              .map(SamplingStatisticsDocument::getRuleName)
-              .collect(Collectors.toSet());
+          statistics.stream().map(SamplingStatisticsDocument::getRuleName).collect(toSet());
 
       GetSamplingTargetsResponse response =
           client.getSamplingTargets(GetSamplingTargetsRequest.create(statistics));
       Map<String, SamplingTargetDocument> targets =
           response.getDocuments().stream()
-              .collect(Collectors.toMap(SamplingTargetDocument::getRuleName, Function.identity()));
+              .collect(toMap(SamplingTargetDocument::getRuleName, identity()));
       updateInternalSamplers(xrayRulesSampler.withTargets(targets, requestedTargetRuleNames, now));
     } catch (Throwable t) {
       // Might be a transient API failure, try again after a default interval.
       fetchTargetsFuture =
-          executor.schedule(
-              this::fetchTargets, DEFAULT_TARGET_INTERVAL_NANOS, TimeUnit.NANOSECONDS);
+          executor.schedule(this::fetchTargets, DEFAULT_TARGET_INTERVAL_NANOS, NANOSECONDS);
       return;
     }
 
     long nextTargetFetchIntervalNanos =
         xrayRulesSampler.nextTargetFetchTimeNanos() - clock.nanoTime();
     fetchTargetsFuture =
-        executor.schedule(this::fetchTargets, nextTargetFetchIntervalNanos, TimeUnit.NANOSECONDS);
+        executor.schedule(this::fetchTargets, nextTargetFetchIntervalNanos, NANOSECONDS);
   }
 
   @Override