+++ azure-resources/src/test/java/io/opentelemetry/contrib/azure/resource/AzureContainersResourceProviderTest.java
@@ -5,6 +5,7 @@
 
 package io.opentelemetry.contrib.azure.resource;
 
+import static io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions.assertThat;
 import static io.opentelemetry.semconv.ServiceAttributes.SERVICE_NAME;
 import static io.opentelemetry.semconv.ServiceAttributes.SERVICE_VERSION;
 import static io.opentelemetry.semconv.incubating.CloudIncubatingAttributes.CLOUD_PLATFORM;
@@ -13,7 +14,6 @@ import static io.opentelemetry.semconv.incubating.ServiceIncubatingAttributes.SE
 
 import com.google.common.collect.ImmutableMap;
 import io.opentelemetry.sdk.testing.assertj.AttributesAssert;
-import io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions;
 import java.util.HashMap;
 import java.util.Map;
 import org.jetbrains.annotations.NotNull;
@@ -50,7 +50,7 @@ class AzureContainersResourceProviderTest {
 
   @NotNull
   private static AttributesAssert createResource(Map<String, String> map) {
-    return OpenTelemetryAssertions.assertThat(
+    return assertThat(
         new AzureContainersResourceProvider(map).createResource(null).getAttributes());
   }
 }