+++ micrometer-meter-provider/src/test/java/io/opentelemetry/contrib/metrics/micrometer/MicrometerMeterTest.java
@@ -31,7 +31,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
-public class MicrometerMeterTest {
+class MicrometerMeterTest {
   SimpleMeterRegistry meterRegistry;
 
   List<Runnable> callbacks;