+++ resource-providers/src/test/java/io/opentelemetry/contrib/resourceproviders/JettyServiceNameDetectorTest.java
@@ -6,8 +6,7 @@
 package io.opentelemetry.contrib.resourceproviders;
 
 import static io.opentelemetry.contrib.resourceproviders.JettyAppServer.parseJettyBase;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.assertj.core.api.Assertions.assertThat;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -15,21 +14,21 @@ import java.nio.file.Path;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
-public class JettyServiceNameDetectorTest {
+class JettyServiceNameDetectorTest {
 
   @Test
   void testJettyBase(@TempDir Path tempDir) throws IOException {
-    assertNull(parseJettyBase(null));
-    assertNull(parseJettyBase(""));
-    assertNull(parseJettyBase("jetty.base="));
-    assertEquals(tempDir.toString(), parseJettyBase("jetty.base=" + tempDir).toString());
-    assertEquals(
-        tempDir.toString(), parseJettyBase("foo jetty.base=" + tempDir + " bar").toString());
+    assertThat(parseJettyBase(null)).isNull();
+    assertThat(parseJettyBase("")).isNull();
+    assertThat(parseJettyBase("jetty.base=")).isNull();
+    assertThat(parseJettyBase("jetty.base=" + tempDir).toString()).isEqualTo(tempDir.toString());
+    assertThat(parseJettyBase("foo jetty.base=" + tempDir + " bar").toString())
+        .isEqualTo(tempDir.toString());
 
     Path otherDir = tempDir.resolve("jetty test");
     Files.createDirectory(otherDir);
-    assertEquals(otherDir.toString(), parseJettyBase("jetty.base=" + otherDir).toString());
-    assertEquals(
-        otherDir.toString(), parseJettyBase("foo jetty.base=" + otherDir + " bar").toString());
+    assertThat(parseJettyBase("jetty.base=" + otherDir).toString()).isEqualTo(otherDir.toString());
+    assertThat(parseJettyBase("foo jetty.base=" + otherDir + " bar").toString())
+        .isEqualTo(otherDir.toString());
   }
 }